pyproject.toml
[tool.poetry]
name = "fastapi-template"
version = "0.1.0"
description = "A template project to create an API with file managment, user authentication and PostgreSQL database."
authors = [
    "Jonathan Bereyziat <jonathan@bereyziat.dev>"
]

[tool.poetry.dependencies]
python = "^3.11.0"
uvicorn = "^0.25.0"
fastapi = "^0.108.0"
pydantic = { version = "^2.4.2", extras = ["email"] }
emails = "^0.6"
gunicorn = "^21.2.0"
psycopg2 = "^2.9.9"
alembic = "^1.13.1"
SQLAlchemy = "^2.0.21"
tenacity = "^8.2.3"
pydantic-settings = "^2.0.3"
python-multipart = "^0.0.6"
sqlalchemy = "^2.0.24"
jinja2 = "^3.1.2"
pyjwt = "^2.8.0"
fastapi-sso = "^0.9.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-cov = "^4.1.0"
flake8 = "^6.1.0"
mypy = "^1.8.0"
httpx = "^0.26.0"
isort = "^5.13.2"
black = "^23.12.1"
autoflake = "^2.2.1"
sqlalchemy-stubs = "^0.4"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 88

